Fix XHCI spec-version check and 64-bit MMIO base truncation
- uefi_verify_xhci(): read HCIVERSION (cap offset 0x00, bits 16:31, BCD)
  instead of HCCPARAMS1 bits 31:24 (MaxPSASize) at cap_len+0x10. The old
  check returned EFI_UNSUPPORTED on virtually all real controllers.
- xhci_verify_version(): apply the same HCIVERSION fix (identical bug).
- Read the full 64-bit BAR0 (offsets 0x10 + 0x14) and widen the MMIO base
  to UINT64 across usb_discovery.c, usb_topology.h, xhci.c, xhci_status.h,
  and the l1_harness debug print, so XHCI MMIO above 4 GB is handled.
- Update test_xhci_fault.c mock to encode the spec version in HCIVERSION.

+++ src/bridge/usb_topology.h
@@ -35,7 +35,7 @@ typedef struct {
 typedef struct {
     USB_ENDPOINT kbd;            /* keyboard interrupt IN endpoint */
     USB_ENDPOINT mouse;          /* mouse interrupt IN endpoint */
-    UINT32       xhci_mmio_base; /* XHCI MMIO base (BAR0), for B1 direct drive */
+    UINT64       xhci_mmio_base; /* XHCI MMIO base (BAR0, 64-bit), for B1 */
     UINT32       xhci_cap_len;   /* CAPLENGTH, for B1 register offsets */
 } USB_TOPOLOGY;
